Make Sprite a class
